+++ packages/contact-center/cc-components/src/index.ts
@@ -1,0 +1,6 @@
+import UserStateComponent from './components/UserState/user-state';
+import {IUserState} from './components/UserState/user-state.types';
+import '@momentum-design/fonts/dist/css/fonts.css';
+import '@momentum-design/tokens/dist/css/components/complete.css';
+
+export {UserStateComponent, type IUserState};