+++ widgets-samples/cc/samples-cc-wc-app/app.js
@@ -1,10 +1,17 @@
 const accessTokenElem = document.getElementById('access_token_elem');
+const themeElem = document.getElementById('theme');
 const widgetsContainer = document.getElementById('widgets-container');
 const ccStationLogin = document.getElementById('cc-station-login');
 const ccUserState = document.createElement('widget-cc-user-state');
 const ccIncomingTask = document.createElement('widget-cc-incoming-task');
 const ccTaskList = document.createElement('widget-cc-task-list');
 const ccCallControl = document.createElement('widget-cc-call-control');
+const themeProviderElem = document.getElementById('theme-provider-elem'); 
+
+themeElem.addEventListener('change', () => {
+    store.setCurrentTheme(themeElem.checked ? 'DARK' : 'LIGHT');
+    themeProviderElem.setAttribute('themeclass', themeElem.checked ? 'mds-theme-stable-darkWebex' : 'mds-theme-stable-lightWebex');
+});
 
 if (!ccStationLogin && !ccUserState) {
     console.error('Failed to find the required elements');